requirements.txt
streamlit
openai
faiss-cpu
sentence-transformers
python-docx
python-pptx
PyPDF2
